pyproject.toml
[project]
name = "wildetect"
version = "0.1.0"
description = "WildDetect - Wildlife Detection System"
readme = "README.md"
requires-python = "~=3.11.0"
dependencies = [
    "fiftyone==1.2.0",
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "huggingface",
    "pillow>=11.3.0",
    "pre-commit>=4.2.0",
    "pyproj>=3.7.1",
    "pytest>=8.4.1",
    "ruff>=0.1.6",
    "shapely>=2.1.1",
    "spyder-kernels==3.0.*",
    "ultralytics>=8.3.162",
    "utm>=0.8.1",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "transformers>=4.53.1",
    "accelerate>=1.8.1",
    "mlflow>=3.1.1",
    "streamlit>=1.46.1",
    "torch==2.6.0",
    "rich>=14.0.0",
    "typer>=0.16.0",
    "albumentations",
    "fiftyone-brain>=0.18.2",
    "torchmetrics>=1.7.4",
    "pip>=25.1.1",
    "label-studio-sdk>=1.0.18",
    "memory-profiler>=0.61.0",
    "line-profiler>=5.0.0",
    "snakeviz>=2.2.2",
    "sahi>=0.11.31",
    "psutil",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "reflex>=0.8.4",
    "litserve>=0.2.15",
    "omegaconf>=2.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "aiohttp>=3.12.15",
    "tenacity>=9.1.2",
    "supervision>=0.26.1",
    "label-studio-tools>=0.0.4",
    "optuna>=4.5.0",
    "rasterio>=1.4.3",
    "slidingwindow>=0.0.14",
    "wildtrain",
    "wildata"
]

[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = ["I", "E", "F", "W", "C90"]
lint.ignore = []

[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[project.scripts]
wildetect = "wildetect.cli:app"

[tool.uv.sources]
wildtrain = { workspace = true }
wildata = { workspace = true }

[tool.uv.workspace]
members = ["wildtrain","wildata"]
